+++ src/core/store-id.ts
@@ -3,6 +3,11 @@ import { sitesStore } from '@src/infrastructure/prun-api/data/sites';
 import { warehousesStore } from '@src/infrastructure/prun-api/data/warehouses';
 import { shipsStore } from '@src/infrastructure/prun-api/data/ships';
 
+export function getBaseStore(planetNaturalIdOrName: string | null | undefined) {
+  const site = sitesStore.getByPlanetNaturalIdOrName(planetNaturalIdOrName);
+  return storagesStore.getByAddressableId(site?.siteId)?.find(x => x.type === 'STORE');
+}
+
 export function getInvStore(invParameter: string | null | undefined) {
   if (!invParameter) {
     return undefined;
@@ -17,9 +22,6 @@ export function getInvStore(invParameter: string | null | undefined) {
     const ship = shipsStore.getByRegistration(invParameter);
     store = storagesStore.getById(ship?.idShipStore);
   }
-  if (!store) {
-    const site = sitesStore.getByPlanetNaturalId(invParameter);
-    store = storagesStore.all.value?.find(x => x.addressableId === site?.siteId);
-  }
+  store ??= getBaseStore(invParameter);
   return store;
 }